docs/views: map urls to views
Document the views in satchmo in docs/views.txt. Right now, it has two
sections: "URLs by Name" and "Views by App".

"URLs by Name" shows the name and view for a url to a view.

"Views by App" provides info on a view, the template path used, and the
context provided to the template.

Add the section "Templates by View" to docs/templates and a link to
docs/views, since the views doc has some info relevant to templates,
such as which url maps to which template, and the context variables
provided.

Right now, we just document the product app.