ContactInfoForm: display a dropdown for state, even if Config.in_country_only is off
Now that we use ajax to re-populate the state field, we don't have to
restrict ourselves to displaying a dropdown only when a user can only be
in one country (ie. Config.in_country_only == True).

The code to create ChoiceFields for state is moved after the country
field creation to take advantage, so that the choices for state fields
follow the initial country determined.