shop/checkout/form.html: re-populate state fields on-the-fly
Use ajax and provide a view to do this.

The JavaScript code needed lives in templates that can be included
easily if needed by satchmo hackers/skinners.

If needed, abide by PAYMENT.COUNTRY_MATCH and follow
_country_match_script by populating both 'state' and 'ship_state'
fields.

Also display a simple message to the user to indicate that state values
are being retrieved. Callbacks are provided to allow for some
extensibility.